fix: 동행 상태에 대한 서비스 인터페이스 추가 및 적용
